Function concordium_smart_contract_engine::v1::resume_receive

pub fn resume_receive<BackingStore: BackingStoreLoad>(
    interrupted_state: Box<ReceiveInterruptedState<CompiledFunction>>, 
    response: InvokeResponse, 
    energy: InterpreterEnergy, 
    state_trie: &mut MutableState, 
    state_updated: bool, 
    backing_store: BackingStore
) -> ExecResult<ReceiveResult<CompiledFunction>>

Resume execution of the receive method after handling the interrupt.

The arguments are as follows

interrupted_state is the state of execution that is captured when we started handling the interrupt

response is the response from the operation that was invoked

energy is the remaning energy left for execution

state_trie is the current state of the contract instance, after handling the interrupt

state_updated indicates whether the state of the instance has changed during handling of the operation. This can currently only happen if there is re-entrancy, i.e., if during handling of the interrupt the instance that invoked it is itself again invoked. Note that this indicates only state changes. Amount changes are no reflected in this.

backing_store gives access to any on-disk storage for the instance state.