Trait concordium_smart_contract_engine::v1::trie::ToSHA256

pub trait ToSHA256<Ctx> {
    fn hash(&self, ctx: &mut Ctx) -> Hash;
}

Compute SHA256 of data given the context. The context is sometimes needed since data is not available in memory and must be first loaded from, e.g., external storage

Required methods

fn hash(&self, ctx: &mut Ctx) -> Hash